+++ .gitignore
@@ -9,3 +9,4 @@ deploy/gen/
 hack/.*
 /*.kubeconfig
 /etc
+/*.yaml
